First pass refactor of Q and A modals.
